+++ packages/google_sign_in/google_sign_in_android/lib/google_sign_in_android.dart
@@ -28,6 +28,11 @@ class GoogleSignInAndroid extends GoogleSignInPlatform {
   }
 
   @override
+  Future<void> clearAuthCache({required String token}) {
+    return _hostApi.clearAuthCache(token);
+  }
+
+  @override
   Future<void> init(InitParameters params) async {
     _hostedDomain = params.hostedDomain;
     _serverClientId =