+++ packages/google_sign_in/google_sign_in_android/pigeons/messages.dart
@@ -194,6 +194,10 @@ abstract class GoogleSignInApi {
   @async
   void clearCredentialState();
 
+  /// Clears the authorization cache for the given token.
+  @async
+  void clearAuthCache(String token);
+
   /// Requests authorization tokens via AuthorizationClient.
   @async
   AuthorizeResult authorize(