+++ packages/google_sign_in/google_sign_in_web/lib/src/gis_client.dart
@@ -279,6 +279,14 @@ class GisSdkClient {
     await signOut();
   }
 
+  /// Clears the authorization cache for the given [token].
+  void clearAuthCache(String token) {
+    _lastClientAuthorizationByUser.removeWhere(
+      (String? key, (TokenResponse tokenResponse, DateTime expiration) value) =>
+          value.tokenResponse.access_token == token,
+    );
+  }
+
   /// Requests the given list of [scopes], and returns the resulting
   /// authorization token if successful.
   ///