+++ packages/google_sign_in/google_sign_in_web/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 1.0.1
 
+* feat(google_sign_in): Add `clearAuthCache` method to remove an access token from the cache.
 * Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 
 ## 1.0.0